+++ src/Control/Concurrent/RCU/GC/Internal.hs
@@ -52,7 +52,7 @@ import Data.List
 import Data.Primitive
 import Prelude hiding (Read(..))
 import System.Mem
-
+import qualified Control.Monad.Fail as Fail
 
 --------------------------------------------------------------------------------
 -- * Shared References
@@ -140,7 +140,10 @@ instance Monad (ReadingRCU s) where
   ReadingRCU m >>= f = ReadingRCU $ \ s -> do
     a <- m s
     runReadingRCU (f a) s
-  fail s = ReadingRCU $ \ _ -> fail s
+  fail = Fail.fail
+
+instance Fail.MonadFail (ReadingRCU s) where
+  fail s = ReadingRCU $ \ _ -> Fail.fail s
 
 instance Alternative (ReadingRCU s) where
   empty = ReadingRCU $ \ _ -> empty
@@ -174,7 +177,10 @@ instance Monad (WritingRCU s) where
   WritingRCU m >>= f = WritingRCU $ \ s -> do
     a <- m s
     runWritingRCU (f a) s
-  fail s = WritingRCU $ \ _ -> fail s
+  fail = Fail.fail
+
+instance Fail.MonadFail (WritingRCU s) where
+  fail s = WritingRCU $ \ _ -> Fail.fail s
 
 instance Alternative (WritingRCU s) where
   empty = WritingRCU $ \ _ -> empty