+++ CHANGELOG.markdown
@@ -1,16 +1,23 @@
+## next [????.??.??]
+--------------------
+* Add `MonadFail` instances for `ReadingRCU` and `WritingRCU`.
+
 ## 0.2.2 [2018.02.06]
 ---------------------
 * Include `HLint.hs` with the tarball distribution, fixing the `hlint`
   test suite.
 
 ## 0.2.1
+--------
 * Support `doctest-0.12`
 
 ## 0.2
+------
 * Revamp `Setup.hs` to use `cabal-doctest`. This makes it build
   with `Cabal-2.0`, and makes the `doctest`s work with `cabal new-build` and
   sandboxes.
 
 ## 0
+----
 * Initial version